Extract verification loading and lead-time panel plotter into shared modules
Move the verif-netCDF loading helpers (_ensure_unique_lead_time,
_select_best_sources, the long-form DataFrame builder, subset_df) into
src/verification/loading.py, and the per-axes metric-vs-lead-time
plotter into src/plotting/metric_lead_time_panel.py. Update
verification_plot_metrics.py and report_experiment_dashboard.py to
import from the new locations; the sys.path hack in the dashboard
script is no longer needed. No behavior change.
